Last commit :fire:

+++ src/main/scala/mx/cinvestav/domain/TraceData.scala
@@ -2,4 +2,4 @@ package mx.cinvestav.domain
 
 import cats.data.Chain
 
-case class TraceData(data:Chain[Int], workers:Int, loadBalancer: Int,basePort:Int)
+case class TraceData(data:Chain[Float], workers:Int, loadBalancer: Int,basePort:Int)
